Remove noex::Table.
Obviously I removed any tests for it too, which is the bulk of this
change. I also made the last adjustments in `bigtable::Table` to not
depend on the `noex::` version, and removed many (but probably not all)
supporting classes for Async* functions. There are separate bugs open to
remove these supporting classes.
