+++ google/cloud/bigtable/bigtable_client.bzl
@@ -41,13 +41,10 @@ bigtable_client_hdrs = [
     "internal/async_op_traits.h",
     "internal/async_poll_op.h",
     "internal/async_read_stream_impl.h",
-    "internal/async_sample_row_keys.h",
-    "internal/async_read_row_operation.h",
     "internal/async_retry_multi_page.h",
     "internal/async_retry_op.h",
     "internal/async_retry_unary_rpc.h",
     "internal/async_retry_unary_rpc_and_poll.h",
-    "internal/async_row_reader.h",
     "internal/bulk_mutator.h",
     "internal/client_options_defaults.h",
     "internal/completion_queue_impl.h",
@@ -61,7 +58,6 @@ bigtable_client_hdrs = [
     "internal/rpc_policy_parameters.h",
     "internal/rowreaderiterator.h",
     "internal/strong_type.h",
-    "internal/table.h",
     "internal/unary_client_utils.h",
     "idempotent_mutation_policy.h",
     "mutation_batcher.h",
@@ -95,7 +91,6 @@ bigtable_client_srcs = [
     "instance_config.cc",
     "instance_update_config.cc",
     "internal/async_bulk_apply.cc",
-    "internal/async_sample_row_keys.cc",
     "internal/bulk_mutator.cc",
     "internal/completion_queue_impl.cc",
     "internal/common_client.cc",
@@ -103,7 +98,6 @@ bigtable_client_srcs = [
     "internal/prefix_range_end.cc",
     "internal/readrowsparser.cc",
     "internal/rowreaderiterator.cc",
-    "internal/table.cc",
     "idempotent_mutation_policy.cc",
     "mutation_batcher.cc",
     "mutations.cc",
