+++ google/cloud/bigtable/bigtable_client_testing.bzl
@@ -18,7 +18,6 @@
 
 bigtable_client_testing_hdrs = [
     "testing/embedded_server_test_fixture.h",
-    "testing/internal_table_test_fixture.h",
     "testing/mock_admin_client.h",
     "testing/mock_data_client.h",
     "testing/mock_instance_admin_client.h",
@@ -35,7 +34,6 @@ bigtable_client_testing_hdrs = [
 
 bigtable_client_testing_srcs = [
     "testing/embedded_server_test_fixture.cc",
-    "testing/internal_table_test_fixture.cc",
     "testing/inprocess_data_client.cc",
     "testing/inprocess_admin_client.cc",
     "testing/table_integration_test.cc",
