+++ google/cloud/bigtable/bigtable_client_unit_tests.bzl
@@ -38,20 +38,11 @@ bigtable_client_unit_tests = [
     "instance_config_test.cc",
     "instance_update_config_test.cc",
     "internal/async_longrunning_op_test.cc",
-    "internal/async_poll_op_test.cc",
     "internal/async_retry_multi_page_test.cc",
-    "internal/async_retry_op_test.cc",
-    "internal/async_retry_unary_rpc_and_poll_test.cc",
     "internal/async_retry_unary_rpc_test.cc",
     "internal/bulk_mutator_test.cc",
-    "internal/table_async_check_and_mutate_row_test.cc",
     "internal/grpc_error_delegate_test.cc",
     "internal/prefix_range_end_test.cc",
-    "internal/table_async_apply_test.cc",
-    "internal/table_async_bulk_apply_test.cc",
-    "internal/table_async_row_reader_test.cc",
-    "internal/table_async_sample_row_keys_test.cc",
-    "internal/table_test.cc",
     "mutation_batcher_test.cc",
     "mutations_test.cc",
     "table_admin_test.cc",
