+++ google/cloud/bigtable/internal/async_bulk_apply.h
@@ -74,84 +74,6 @@ class AsyncRetryBulkApply
   promise<std::vector<FailedMutation>> promise_;
 };
 
-/**
- * Perform an AsyncBulkApply operation request with retries.
- *
- * @tparam Client the class implementing the asynchronous operation, examples
- *     include `DataClient`, `AdminClient`, and `InstanceAdminClient`.
- *
- * @tparam MemberFunctionType the type of the member function to call on the
- *     `Client` object. This type must meet the requirements of
- *     `internal::CheckAsyncUnaryRpcSignature`, the `AsyncRetryUnaryRpc`
- *     template is disabled otherwise.
- *
- * @tparam Functor the type of the function-like object that will receive the
- *     results. It must satisfy (using C++17 types):
- *     static_assert(std::is_invocable_v<
- *         Functor, CompletionQueue&, std::vector<FailedMutation>&,
- *             grpc::Status&>);
- *
- * @tparam valid_callback_type a format parameter, uses `std::enable_if<>` to
- *     disable this template if the functor does not match the expected
- *     signature.
- */
-template <typename Functor,
-          typename std::enable_if<
-              google::cloud::internal::is_invocable<
-                  Functor, CompletionQueue&, std::vector<FailedMutation>&,
-                  grpc::Status&>::value,
-              int>::type valid_callback_type = 0>
-class AsyncRetryBulkApplyNoex
-    : public AsyncRetryOp<ConstantIdempotencyPolicy, Functor,
-                          AsyncBulkMutatorNoex> {
- public:
-  using MutationsSucceededFunctor =
-      AsyncBulkMutatorNoex::MutationsSucceededFunctor;
-  using MutationsFailedFunctor = AsyncBulkMutatorNoex::MutationsFailedFunctor;
-  using AttemptFinishedFunctor = AsyncBulkMutatorNoex::AttemptFinishedFunctor;
-
-  AsyncRetryBulkApplyNoex(std::unique_ptr<RPCRetryPolicy> rpc_retry_policy,
-                          std::unique_ptr<RPCBackoffPolicy> rpc_backoff_policy,
-                          IdempotentMutationPolicy& idempotent_policy,
-                          MetadataUpdatePolicy metadata_update_policy,
-                          std::shared_ptr<bigtable::DataClient> client,
-                          bigtable::AppProfileId const& app_profile_id,
-                          bigtable::TableId const& table_name, BulkMutation mut,
-                          Functor callback)
-      : AsyncRetryOp<ConstantIdempotencyPolicy, Functor, AsyncBulkMutatorNoex>(
-            __func__, std::move(rpc_retry_policy),
-            // BulkMutator is idempotent because it keeps track of idempotency
-            // of the mutations it holds.
-            std::move(rpc_backoff_policy), ConstantIdempotencyPolicy(true),
-            std::move(metadata_update_policy), std::move(callback),
-            AsyncBulkMutatorNoex(client, std::move(app_profile_id),
-                                 std::move(table_name), idempotent_policy,
-                                 std::move(mut))) {}
-
-  AsyncRetryBulkApplyNoex(
-      std::unique_ptr<RPCRetryPolicy> rpc_retry_policy,
-      std::unique_ptr<RPCBackoffPolicy> rpc_backoff_policy,
-      IdempotentMutationPolicy& idempotent_policy,
-      MetadataUpdatePolicy metadata_update_policy,
-      std::shared_ptr<bigtable::DataClient> client,
-      bigtable::AppProfileId const& app_profile_id,
-      bigtable::TableId const& table_name, BulkMutation mut,
-      MutationsSucceededFunctor mutations_succeeded_callback,
-      MutationsFailedFunctor mutations_failed_callback,
-      AttemptFinishedFunctor attempt_finished_callback, Functor callback)
-      : AsyncRetryOp<ConstantIdempotencyPolicy, Functor, AsyncBulkMutatorNoex>(
-            __func__, std::move(rpc_retry_policy),
-            // BulkMutator is idempotent because it keeps track of idempotency
-            // of the mutations it holds.
-            std::move(rpc_backoff_policy), ConstantIdempotencyPolicy(true),
-            std::move(metadata_update_policy), std::move(callback),
-            AsyncBulkMutatorNoex(
-                client, std::move(app_profile_id), std::move(table_name),
-                idempotent_policy, std::move(mutations_succeeded_callback),
-                std::move(mutations_failed_callback),
-                std::move(attempt_finished_callback), std::move(mut))) {}
-};
-
 }  // namespace internal
 }  // namespace BIGTABLE_CLIENT_NS
 }  // namespace bigtable
