+++ google/cloud/bigtable/tests/bigtable_client_integration_tests.bzl
@@ -20,7 +20,6 @@ bigtable_client_integration_tests = [
     "admin_integration_test.cc",
     "admin_async_future_integration_test.cc",
     "data_async_future_integration_test.cc",
-    "data_async_integration_test.cc",
     "data_integration_test.cc",
     "filters_integration_test.cc",
     "instance_admin_integration_test.cc",
